Fix lwIP IPv4 routing logic and add ARP resolution to EkaTcpSess::connect.
Routing and ARP resolution did not work properly prior to this commit.
The implementation worked by adding a static ARP entry directly for
the destination IP, claiming its MAC was the gateway's MAC address.
Until recently, the netmask and gateway were not being set at the
netif level. This made lwIP believe that the destination was always
on the same network, so the routing was trivial. It was also able to
find the hacky ARP entries that associated all destinations with the
gateway's MAC.

Once the netif mask and gateway started being set, the lwIP stack was
able to "realize" that the destination was not on the same network,
and tried to do actual routing. This always failed with EHOSTUNREACH,
"no route to host".

This commit does a few things:

- Added an lwIP hook to help with source-based routing. This new
  function (eka_route_ipv4_src) does not select the final next hop
  address but only decides which netif to use. Once lwIP is told the
  correct netif, it will compute the next hop to be the destination
  IP or the gateway IP, depending on the netif's current netmask and
  gateway settings. The provided function returns the netif associated
  with the source IP.
- Added a global list of all open EkaDev instances and an rwlock to
  protect it. This is needed to support the above, because the routing
  function is only given the source and destination IP addresses, and
  needs to look through all the open Ekaline cores to find the the one
  associated with that source IP address, and then return its
  pLwipNetIf.
- Added a function (ekaAddArpEntry) to add static ARP entries
  rather than doing this in EkaTcpSess::connect. This is called
  from EkaDev::configurePort when the user directy provides the
  gateway MAC.
- EkaTcpSess::connect now does it's own version of the same next-hop
  routing logic that lwIP will do, so it can set EkaTcpSess::macDa
  prior to lwip_connect being called. It starts by looking in the
  ARP cache, since the MAC of the next hop may already be present
  from a previous connect or from a static ARP entry added via
  `ekaAddArpEntry`. If no entry is found, it sends an ARP request as
  before, but this time for the proper next-hop IP address rather
  than for the destination.